backend/requirements.txt
argcomplete==3.6.2
asgiref==3.8.1
boto3==1.37.30
botocore==1.37.30
certifi==2025.1.31
cfn-flip==1.3.0
charset-normalizer==3.4.1
click==8.1.8
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
durationpy==0.9
hjson==3.1.0
idna==3.10
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==3.0.2
placebo==0.9.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
troposphere==4.9.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
zappa==0.59.0
